chore: break out node/index into multiple files
